storage: Fix build on i686
off_t is signed and it's size is the same as long only on 64b archs.
Thus it cannot be formatted as %lu.

+++ src/storage/storage_backend_logical.c
@@ -116,9 +116,9 @@ virStorageBackendLogicalInitializeDevice(const char *path)
 
     if (size < 4 * PV_BLANK_SECTOR_SIZE) {
         virReportError(VIR_ERR_OPERATION_UNSUPPORTED,
-                       _("cannot initialize '%s' detected size='%lu' less "
+                       _("cannot initialize '%s' detected size='%zd' less "
                          "than minimum required='%d"),
-                         path, size, 4 * PV_BLANK_SECTOR_SIZE);
+                         path, (ssize_t) size, 4 * PV_BLANK_SECTOR_SIZE);
         goto cleanup;
     }
     if ((size = lseek(fd, 0, SEEK_SET)) == (off_t)-1) {
